Dataset to be used for this project.
We use a Network backup Dataset, which is comprised of simulated traffic data on a
backup system over a network. The system monitors the files residing in a destination
machine and copies their changes in four hour cycles. At the end of each backup process,
the size of the data moved to the destination as well as the duration it took are logged,
to be used for developing prediction models. We dene a work
ow as a task that backs
up data from a group of files, which have similar patterns of change in terms of size over
time. The dataset has around 18000 data points with the following columns/variables:
* Week index
* Day of the week at which the file back up has started
* Backup start time: Hour of the day
* Workflow ID
* File name
* Backup size: the size of the file that is backed up in that cycle in GB
* Backup time: the duration of the backup procedure in hour